+++ src/Core/NESBoard.h
@@ -1,8 +1,9 @@
 #pragma once
 #include "Core/Bus/Bus.h"
 #include "Core/CPU/CPU.h"
+#include "Core/PPU/PPU.h"
 #include "Core/Memory/RAM/RAM.h"
-#include "Core/Cartridge/Cartridge.h" // Adicionado para incluir o Cartridge
+#include "Core/Cartridge/Cartridge.h"
 #include <string>
 
 namespace R2NES::Core
@@ -18,7 +19,9 @@ namespace R2NES::Core
 
         void reset();
 
-        CPU& getCpu() { return cpu; } // Retorna por referência para permitir modificação (L-value)
+        CPU& getCpu() { return cpu; }
+
+        PPU& getPpu() { return ppu; }
 
         // Permite ler um byte de qualquer lugar do barramento (debug)
         uint8_t cpuRead(uint16_t addr) { return bus.cpuRead(addr); }
@@ -30,7 +33,7 @@ namespace R2NES::Core
         Bus bus;
         RAM ram;
         CPU cpu;
-        // Ppu ppu; // Será adicionado mais tarde
+        PPU ppu;
 
         // Controle de timing
         uint32_t systemClockCounter = 0;